JP-BE-PY-batch-2/class_25/blogging_app/requirements.txt
djangorestframework-simplejwt
drf-yasg
django-jazzmin